requirements.txt
python-telegram-bot==20.4
aiohttp==3.9.4



